refactor : 공통 컴포넌트 Commons 폴더로 이동
